+++ backend/game/player.py
@@ -51,7 +51,6 @@ class Player:
         self.cards = None
 
         self.current_task = None
-        self.is_active = False
 
         self.card_lock = threading.Lock()
         self.task_lock = threading.Lock()
@@ -79,11 +78,6 @@ class Player:
         with self.card_lock:
             self.cards[new_card.id] = new_card
             return self.cards.pop(card_to_replace_id)
-
-    def toggle_is_active(self):
-        self.is_active = not self.is_active
-        self.__update_state()
-
 
     def call_tricks(self, called_tricks: int, max_tricks: int, is_last: bool) -> int:
         logging.debug("Calling call_tricks for " + self.name)
@@ -171,7 +165,7 @@ class Player:
 
     def __update_state(self):
         with self.state_lock:
-            self.state = PlayerState(self.user, self.score, self.is_active, self.tricks_called, self.tricks_made)
+            self.state = PlayerState(self.user, self.score, self.current_task is not None, self.tricks_called, self.tricks_made)
 
     def __get_hand_card(self, card: Card, lead_color: str=None):
         playable_cards = self.__get_playable_cards(lead_color)
@@ -192,7 +186,6 @@ class PlayerTask:
     def __init__(self, task_type: str, player: Player, options: List[str]):
         self.task_type = task_type
         self.player = player
-        self.player.toggle_is_active()
 
         self.options = options
         self.selected = None
@@ -224,6 +217,5 @@ class PlayerTask:
             self.selected = user_input
 
             self.player.current_task = None
-            self.player.toggle_is_active()
 
             self.input_event.set()